# Conflicts:
#	src/main/java/projetoSoftware/com/mercado/controller/ClienteController.java

+++ src/main/java/projetoSoftware/com/mercado/repository/ReclamacaoRepository.java
@@ -3,6 +3,6 @@ package projetoSoftware.com.mercado.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import projetoSoftware.com.mercado.model.Reclamacao;
 
-public interface ReclamacaoRepository extends JpaRepository<Reclamacao, Long> {
+public interface ReclamacaoRepository extends JpaRepository<Reclamacao, Integer> {
     Reclamacao save(Reclamacao reclamacao);
 }
